+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Allow attributes on enum values. [#38](https://github.com/vlang/intellij-v/issues/38)
 - Fix a freeze after clicking button to add a toolchain. [#37](https://github.com/vlang/intellij-v/issues/37)
 - Allow newlines in between operators in expressions. [#34](https://github.com/vlang/intellij-v/issues/34)
+- Introduce links in console when program panic's and shows backtrace. [#12](https://github.com/vlang/intellij-v/issues/12)
 
 ## [2025.2.1] - 2025-08-09